refactor: move validator out and extend docs
