+++ packages/core/src/shared-wallets/docs/schema/README.md
@@ -5,21 +5,74 @@ This directory contains JSON schemas and example files for the shared wallet str
 ## Schemas
 
 ### `shared-wallet-schema.json`
+
 This schema defines the structure of a shared wallet. It includes metadata about the wallet and a reference to the native script. The metadata contains the shared wallet name and a list of participants, each with a name and an Ed25519KeyHash public key.
 
 ### `cardano-babbage.json`
+
 This schema defines various types used in Cardano transactions, including definitions for native scripts, addresses, and other related types.
 
 ## Examples
 
 ### `shared-wallet-scriptAll-example.json`
+
 This example demonstrates a shared wallet using the `ScriptAll` native script type, which requires all specified scripts to validate for the transaction to be considered valid.
 
 ### `shared-wallet-scriptAny-example.json`
+
 This example demonstrates a shared wallet using the `ScriptAny` native script type, which requires at least one of the specified scripts to validate for the transaction to be considered valid.
 
 ### `shared-wallet-scriptNOfK-example.json`
+
 This example demonstrates a shared wallet using the `ScriptNOfK` native script type, which requires a minimum number (`N`) of the specified scripts to validate for the transaction to be considered valid.
+
+## Schema Validation with Zod
+
+### Running Your JSON Schema Against a Validator Sandbox
+
+You can run your schema against a validator online using https://zod-playground.vercel.app/.
+
+### Your Reference for Comprehensive Zod Documentation
+
+How to validate, information on types and inference can be found on https://zod.dev/.
+
+### Generating a Zod Schema Online
+
+You can generate a Zod compatible schema from a JSON schema online using https://stefanterdell.github.io/json-schema-to-zod-react/.
+
+### Zod Usage Example
+
+The following code snippets show how to validate using Zod.
+
+#### Raw Parsing
+
+```
+try {
+  const result = sharedWalletJsonSchema.parse(<my json schema>);
+  console.log("Schema is valid.")
+} catch (e) {
+  console.error("Validation error:", e.errors);
+}
+```
+
+#### Safe Parsing (Verbose Errors)
+
+Safe parsing does not throw errors, but rather logs them as part of the returned object. It is preferrable when you need to know what went wrong in the validation.
+
+```
+try {
+  const result = sharedWalletJsonSchema.safeParse(<my json schema>);
+
+  if (result.errors) {
+    console.log("Schema has errors: ", errors);
+  }
+
+  console.log("Schema is valid.");
+  return result.data;
+} catch (e) {
+  console.error("Other potential errors:", e.errors);
+}
+```
 
 ## Regenerating TypeScript Types
 
@@ -27,3 +80,4 @@ To regenerate the TypeScript types from the JSON schema, use the following comma
 
 ```bash
 npx --package json-schema-to-typescript -c 'json2ts shared-wallet-schema.json > shared-wallet-types-autogenerated.ts'
+```
