+++ packages/core/src/shared-wallets/docs/schema/shared-wallet-scriptNOfK-example.json
@@ -4,15 +4,15 @@
     "coSigners": [
       {
         "name": "Alice",
-        "publicKey": "e0a8a80a28f205b69029ac8448265da2cef6e2c70d6a91fcd699ea859df45860f6788ff36f7ed44ccf564974ad2c033e640a5f25b943308935d7103612ef2c1f"
+        "sharedWalletKey": "e0a8a80a28f205b69029ac8448265da2cef6e2c70d6a91fcd699ea859df45860f6788ff36f7ed44ccf564974ad2c033e640a5f25b943308935d7103612ef2c1f"
       },
       {
         "name": "Bob",
-        "publicKey": "e0a8a80a28f205b69029ac8448265da2cef6e2c70d6a91fcd699ea859df45860f6788ff36f7ed44ccf564974ad2c033e640a5f25b943308935d7103612ef2c1f"
+        "sharedWalletKey": "e0a8a80a28f205b69029ac8448265da2cef6e2c70d6a91fcd699ea859df45860f6788ff36f7ed44ccf564974ad2c033e640a5f25b943308935d7103612ef2c1f"
       },
       {
         "name": "Charlie",
-        "publicKey": "b14c6e042df1d47e0d6d76559b8587de821ec96727f372e7d809d8489788f551a71c5dab280ae3440b7965f6455bf14d734eab82ab2ab8e2fe803374fd9db8df"
+        "sharedWalletKey": "b14c6e042df1d47e0d6d76559b8587de821ec96727f372e7d809d8489788f551a71c5dab280ae3440b7965f6455bf14d734eab82ab2ab8e2fe803374fd9db8df"
       }
     ]
   },
