+++ packages/core/src/shared-wallets/utils/index.ts
@@ -4,3 +4,4 @@ export * from './is-shared-wallet';
 export * from './get-shared-wallet-sign-policy';
 export * from './get-key-hash-to-wallet-name-map';
 export * from './has-signed';
+export * from './validate-schema';
